% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Empty Function Cache}
\usage{
clear_cache()
}
\description{
This function will reset the cache for any and all ffscrapr cached functions,
as well as nflreadr's cached functions.
}
