% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_helpers.R
\name{.add_allplay}
\alias{.add_allplay}
\title{Add allplay from a standardised schedule output}
\usage{
.add_allplay(schedule)
}
\arguments{
\item{schedule}{\itemize{
\item an output from ff_schedule
}}
}
\description{
Add allplay from a standardised schedule output
}
\keyword{internal}
