% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn__helpers.R
\name{.espn_activity_map}
\alias{.espn_activity_map}
\title{ESPN Activity/Transaction Mapping}
\usage{
.espn_activity_map()
}
\description{
ESPN Activity/Transaction Mapping
}
\seealso{
\url{https://github.com/cwendt94/espn-api/blob/master/espn_api/football/constant.py#L82-92}
}
\keyword{internal}
