% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn__helpers.R
\name{.espn_lineupslot_map}
\alias{.espn_lineupslot_map}
\title{ESPN Lineup Slot map}
\usage{
.espn_lineupslot_map()
}
\description{
This is for the starting lineup specifically - primary positions is accessible via \code{.espn_pos_map}
}
\seealso{
\url{https://github.com/cwendt94/espn-api/blob/master/espn_api/football/constant.py}
}
\keyword{internal}
