% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn__helpers.R
\name{.espn_pos_map}
\alias{.espn_pos_map}
\title{ESPN Primary Position map}
\usage{
.espn_pos_map()
}
\description{
Decoded by hand - if you have an IDP ESPN league please open a GitHub issue
and pass along the league info so we can expand this.
}
\keyword{internal}
