% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn__helpers.R
\name{.espn_stat_map}
\alias{.espn_stat_map}
\title{ESPN Stat ID map}
\usage{
.espn_stat_map()
}
\description{
ESPN Stat ID map
}
\seealso{
\url{https://github.com/cwendt94/espn-api/blob/master/espn_api/football/constant.py}
}
\keyword{internal}
