% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn__helpers.R
\name{.espn_team_map}
\alias{.espn_team_map}
\title{ESPN Team ID map}
\usage{
.espn_team_map()
}
\description{
Opinionatedly conforming to DynastyProcess standards, which match to MyFantasyLeague.
Abbreviations are consistently three letters.
}
\seealso{
\url{https://github.com/cwendt94/espn-api/blob/master/espn_api/football/constant.py}
}
\keyword{internal}
