% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_helpers.R
\name{.fn_choose_season}
\alias{.fn_choose_season}
\title{Choose current season}
\usage{
.fn_choose_season(date = NULL)
}
\description{
A helper function to return the current year if March or later, otherwise assume previous year
}
\keyword{internal}
