% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_helpers.R
\name{.fn_set_ratelimit}
\alias{.fn_set_ratelimit}
\title{Set rate limit}
\usage{
.fn_set_ratelimit(toggle = TRUE, platform, rate_number, rate_seconds)
}
\arguments{
\item{toggle}{a logical to turn on rate_limiting if TRUE and off if FALSE}

\item{rate_number}{number of calls per \code{rate_seconds}}

\item{rate_seconds}{number of seconds}
}
\description{
A helper function that creates a new copy of the httr::GET function and stores it
in the .ffscrapr_env hidden object
}
\keyword{internal}
