% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_helpers.R
\name{.fn_set_useragent}
\alias{.fn_set_useragent}
\title{Set user agent}
\usage{
.fn_set_useragent(user_agent)
}
\description{
Self-identifying is mostly about being polite, although MFL has a program to give verified clients more bandwidth!
See: https://www03.myfantasyleague.com/2020/csetup?C=APICLI
}
\keyword{internal}
