% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfl_connect.R
\name{.mfl_logincookie}
\alias{.mfl_logincookie}
\title{Get MFL Login Cookie}
\usage{
.mfl_logincookie(user_name, password, season)
}
\arguments{
\item{user_name}{MFL user_name (as string)}

\item{password}{MFL password (as string)}

\item{season}{Season}
}
\value{
a login cookie, which should be included as a parameter in an httr GET request
}
\description{
Gets login cookie for MFL based on user_name/password
Docs: https://api.myfantasyleague.com/2020/api_info#login_info
}
\keyword{internal}
