% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_helpers.R
\name{.retry_get}
\alias{.retry_get}
\title{Create RETRY version of GET}
\usage{
.retry_get(...)
}
\arguments{
\item{...}{arguments passed to \code{httr::GET}}
}
\description{
This wrapper on httr retries the httr::GET function based on best-practice heuristics
}
\keyword{internal}
