% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_helpers.R
\name{.retry_post}
\alias{.retry_post}
\title{Create RETRY version of POST}
\usage{
.retry_post(...)
}
\arguments{
\item{...}{arguments passed to \code{httr::POST}}
}
\description{
This wrapper on httr retries the httr::POST function based on best-practice heuristics.
}
\keyword{internal}
