% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper_connect.R
\name{.sleeper_userid}
\alias{.sleeper_userid}
\title{Get Sleeper User ID}
\usage{
.sleeper_userid(user_name)
}
\arguments{
\item{user_name}{Sleeper username}
}
\value{
sleeper userID
}
\description{
Docs: https://docs.sleeper.app
}
\keyword{internal}
