% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_import_dp.R
\name{dp_clean_html}
\alias{dp_clean_html}
\title{Remove HTML from string}
\usage{
dp_clean_html(names)
}
\arguments{
\item{names}{a character (or character vector)}
}
\value{
a character vector of cleaned strings
}
\description{
Applies some regex to clean html tags from strings. This is useful for platforms such as MFL that interpret HTML in their franchise name fields.
}
\examples{

c(
  "<b><font color= Cyan>Kevin OBrien (@kevinobrienff) </FONT></B>",
  "<em><font color= Purple> Other fun names</font></em>"
) \%>\% dp_clean_html()
}
