% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_deprecated.R
\name{dp_clean_names}
\alias{dp_clean_names}
\alias{dp_cleannames}
\title{Clean Player Names}
\usage{
dp_clean_names(...)

dp_cleannames(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[nflreadr:clean_player_names]{nflreadr::clean_player_names}}
  \describe{
    \item{\code{player_name}}{a character vector of player names}
    \item{\code{lowercase}}{defaults to FALSE - if TRUE, converts to lowercase}
    \item{\code{convert_lastfirst}}{defaults to TRUE - converts names from "Last, First" to "First Last"}
    \item{\code{use_name_database}}{uses internal name database to do common substitutions (Mitchell Trubisky to Mitch Trubisky etc)}
    \item{\code{convert_to_ascii}}{If \code{TRUE}, will transliterate to latin-ascii via the
stringi package. Defaults to \code{TRUE} if the stringi package is installed.}
  }}
}
\value{
a character vector of cleaned names
}
\description{
Applies some name-cleaning heuristics to facilitate joins. These heuristics may include:
\itemize{
\item removing periods and apostrophes
\item removing common suffixes, such as Jr, Sr, II, III, IV
\item converting to lowercase
\item using \code{ffscrapr::dp_name_mapping} to do common name substitutions, such as Mitch Trubisky to Mitchell Trubisky
}
}
\details{
Equivalent to the operation done by \code{ffscrapr::dp_clean_names()} and uses the same player name database.
}
\examples{
\donttest{

dp_cleannames(c("A.J. Green", "Odell Beckham Jr.", "Le'Veon Bell Sr."))

dp_cleannames(c("Trubisky, Mitch", "Atwell, Chatarius", "Elliott, Zeke", "Elijah Moore"),
  convert_lastfirst = TRUE,
  use_name_database = TRUE
)
}
}
\seealso{
\code{nflreadr::clean_player_names()}
}
