% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_import_dp.R
\name{dp_cleannames}
\alias{dp_cleannames}
\title{Clean Names}
\usage{
dp_cleannames(player_name, lowercase = FALSE)
}
\arguments{
\item{player_name}{a character (or character vector)}

\item{lowercase}{defaults to FALSE - if TRUE, converts to lowercase}
}
\value{
a character vector of cleaned names
}
\description{
Applies some name-cleaning heuristics to facilitate joins.
May eventually refer to a name-cleaning database, for now will just include basic regex.
}
\examples{
\donttest{
dp_cleannames(c("A.J. Green", "Odell Beckham Jr.", "Le'Veon Bell Sr."))
}

}
