% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_import_dp.R
\docType{data}
\name{dp_name_mapping}
\alias{dp_name_mapping}
\title{Alternate name mappings}
\format{
A named character vector
\describe{
\item{name attribute}{The name of each element of the vector is the "alternate" name}
\item{value attribute}{The value of each element of the vector is the "correct" name.}
}
}
\usage{
dp_name_mapping
}
\description{
A named character vector mapping common alternate names
}
\keyword{datasets}
