% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_import.R
\name{dp_playerids}
\alias{dp_playerids}
\title{Import latest DynastyProcess player IDs}
\usage{
dp_playerids()
}
\value{
a tibble of trade values
}
\description{
Fetches a copy of the latest DynastyProcess player IDs csv
}
\examples{
dp_playerids()
}
\seealso{
\url{https://github.com/DynastyProcess/data}
}
