% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_api.R
\name{espn_getendpoint}
\alias{espn_getendpoint}
\title{GET any ESPN endpoint}
\usage{
espn_getendpoint(conn, endpoint, ...)
}
\arguments{
\item{conn}{a connection object created by \code{espn_connect} or \code{ff_connect()}}

\item{endpoint}{a string defining which endpoint to return from the API}

\item{...}{Arguments which will be passed as "argumentname = argument" in an HTTP query parameter}
}
\value{
A list object containing the query, response, and parsed content.
}
\description{
The endpoint names and HTTP parameters (i.e. argument names) are CASE SENSITIVE.
Best URL help page is TBD, possibly might just be the vignette.
}
\details{
Check out the vignette for more details and example usage.
}
\seealso{
\code{vignette("espn_getendpoint")}
}
