% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_players.R
\name{espn_players}
\alias{espn_players}
\title{ESPN players library}
\usage{
espn_players(conn, limit = 5000)
}
\arguments{
\item{conn}{a connection object created by \code{espn_connect} or \code{ff_connect()}}

\item{limit}{the maximum number of results to return}
}
\value{
a dataframe containing all ~7000+ players in the ESPN database
}
\description{
A cached table of ESPN NFL players. Will store in memory for each session!
(via memoise in zzz.R)
}
\examples{
\donttest{
  conn <- espn_connect(season = 2020,league_id = 899513)

  player_list <- espn_players(conn)
}

}
