% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mfl_franchises.R,
%   R/sleeper_franchises.R
\name{ff_franchises}
\alias{ff_franchises}
\alias{ff_franchises.mfl_conn}
\alias{ff_franchises.sleeper_conn}
\title{Get League Franchises}
\usage{
ff_franchises(conn)

\method{ff_franchises}{mfl_conn}(conn)

\method{ff_franchises}{sleeper_conn}(conn)
}
\arguments{
\item{conn}{a conn object created by \code{ff_connect()}}
}
\value{
A tibble of franchises, complete with IDs
}
\description{
Returns a tidy dataframe of franchises and any relevant information

Returns a tidy dataframe of franchises and any relevant information
}
\examples{
ssb_conn <- ff_connect(platform = "mfl", league_id = 54040, season = 2020)
ff_franchises(ssb_conn)
jml_conn <- ff_connect(platform = "sleeper", league_id = 522458773317046272, season = 2020)
ff_franchises(jml_conn)
}
