% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfl_leaguesummary.R
\name{ff_league.mfl_conn}
\alias{ff_league.mfl_conn}
\title{Get a summary of common league settings}
\usage{
\method{ff_league}{mfl_conn}(conn)
}
\arguments{
\item{conn}{the connection object created by \code{mfl_connect()}}
}
\value{
A tibble of length one.
}
\description{
This function returns a tibble of common league settings - things like "1QB" or "2QB", best ball, team count etc
}
\seealso{
\url{https://api.myfantasyleague.com/2020/api_info?STATE=details}
}
