% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ff_league}
\alias{ff_league}
\title{Get League Summary}
\usage{
ff_league(conn)
}
\arguments{
\item{conn}{a conn object created by \code{ff_connect()}}
}
\value{
a dataframe of stuff
}
\description{
This function returns a dataframe summarizing common league settings.
}
