% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfl_rosters.R
\name{ff_rosters.mfl_conn}
\alias{ff_rosters.mfl_conn}
\title{Get a dataframe of scoring settings, referencing the "all rules" library endpoint.
The all-rules endpoint is saved to a cache, so subsequent function calls should be faster!}
\usage{
\method{ff_rosters}{mfl_conn}(conn)
}
\arguments{
\item{conn}{a conn object created by \code{ff_connect()}}
}
\description{
Get a dataframe of scoring settings, referencing the "all rules" library endpoint.
The all-rules endpoint is saved to a cache, so subsequent function calls should be faster!
}
\examples{
ssb_conn <- ff_connect(platform = "mfl",league_id = 54040,season = 2020)
ff_rosters(ssb_conn)


}
\seealso{
\url{http://www03.myfantasyleague.com/2020/scoring_rules#rules}
}
