% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mfl_schedule.R,
%   R/sleeper_schedule.R
\name{ff_schedule}
\alias{ff_schedule}
\alias{ff_schedule.mfl_conn}
\alias{ff_schedule.sleeper_conn}
\title{Get Schedule}
\usage{
ff_schedule(conn, ...)

\method{ff_schedule}{mfl_conn}(conn, ...)

\method{ff_schedule}{sleeper_conn}(conn, ...)
}
\arguments{
\item{conn}{a conn object created by \code{ff_connect()}}

\item{...}{additional args which might be used eventually}
}
\value{
A tidy dataframe with one row per game per franchise per week
}
\description{
This function returns a tibble with one row for every team for every weekly matchup
}
\examples{
ssb_conn <- ff_connect(platform = "mfl", league_id = 54040, season = 2020)
ff_schedule(ssb_conn)


jml_conn <- ff_connect(platform = "sleeper", league_id = 522458773317046272, season = 2020)
ff_schedule(jml_conn)


}
