% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mfl_scoring.R
\name{ff_scoring}
\alias{ff_scoring}
\alias{ff_scoring.default}
\alias{ff_scoring.mfl_conn}
\title{Get League Scoring settings}
\usage{
ff_scoring(conn)

ff_scoring.default(conn)

ff_scoring.mfl_conn(conn)
}
\arguments{
\item{conn}{a conn object created by \code{ff_connect()}}
}
\value{
A tibble of league scoring rules for each position defined.
}
\description{
This function returns a dataframe with detailed scoring settings, broken down by position, event, range, and points.
}
\examples{
ssb_conn <- ff_connect(platform = "mfl",league_id = 54040,season = 2020)
ff_scoring(ssb_conn)


}
\seealso{
\url{http://www03.myfantasyleague.com/2020/scoring_rules#rules}
}
