% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_generics.R, R/mfl_scoringhistory.R
\name{ff_scoringhistory}
\alias{ff_scoringhistory}
\alias{ff_scoringhistory.mfl_conn}
\title{Get a dataframe of scoring history, utilizing the ff_scoring and load_player_stats functions.}
\usage{
ff_scoringhistory(conn, season, ...)

\method{ff_scoringhistory}{mfl_conn}(conn, season = 1999:2020, ...)
}
\arguments{
\item{conn}{a conn object created by \code{ff_connect()}}

\item{season}{season a numeric vector of seasons (earliest available year is 1999)}

\item{...}{other arguments}
}
\value{
A tidy dataframe of scoring history in a flat table, one row per player per week.
}
\description{
Get a dataframe of scoring history, utilizing the ff_scoring and load_player_stats functions.

Get a dataframe of scoring history, utilizing the ff_scoring and load_player_stats functions.
}
\section{Methods (by class)}{
\itemize{
\item \code{mfl_conn}: MFL: returns scoring history in a flat table, one row per player per week.
}}

\examples{
\donttest{
}

}
\seealso{
\url{https://www.nflfastr.com/articles/nflfastR.html#yards-from-scrimmage}
}
