% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_generics.R, R/espn_starterpositions.R,
%   R/flea_starterpositions.R, R/mfl_starterpositions.R,
%   R/sleeper_starterpositions.R
\name{ff_starter_positions}
\alias{ff_starter_positions}
\alias{ff_starter_positions.espn_conn}
\alias{ff_starter_positions.flea_conn}
\alias{ff_starter_positions.mfl_conn}
\alias{ff_starter_positions.sleeper_conn}
\title{Get Starting Lineup Settings}
\usage{
ff_starter_positions(conn, ...)

\method{ff_starter_positions}{espn_conn}(conn, ...)

\method{ff_starter_positions}{flea_conn}(conn, ...)

\method{ff_starter_positions}{mfl_conn}(conn, ...)

\method{ff_starter_positions}{sleeper_conn}(conn, ...)
}
\arguments{
\item{conn}{the list object created by \code{ff_connect()}}

\item{...}{other arguments (currently unused)}
}
\value{
A tidy dataframe with every player for every week, including a flag for whether they were started or not
}
\description{
This function returns a tidy dataframe with positional lineup rules.
}
\section{Methods (by class)}{
\itemize{
\item \code{espn_conn}: ESPN: returns who was started as well as what they scored.

\item \code{flea_conn}: Fleaflicker: returns who was started as well as what they scored.

\item \code{mfl_conn}: MFL: returns the player fantasy scores for each week (not the actual stats)

\item \code{sleeper_conn}: Sleeper: returns only "who" was started, without any scoring/stats data. Only returns season specified in initial connection object.
}}

\examples{
\donttest{
conn <- espn_connect(season = 2020, league_id = 1178049)
ff_starter_positions(conn)
}

\donttest{
conn <- fleaflicker_connect(season = 2020, league_id = 206154)
ff_starter_positions(conn)
}

\donttest{
dlfidp_conn <- mfl_connect(2020, league_id = 54040)
ff_starter_positions(conn = dlfidp_conn)
}

\donttest{
test_conn <- sleeper_connect(league_id = "653543448376320000", season = 2021)
test_starters <- ff_starter_positions(jml_conn)
}

}
