% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ff_starters}
\alias{ff_starters}
\title{Get Starting Lineups}
\usage{
ff_starters(conn, ...)
}
\arguments{
\item{conn}{a conn object created by \code{ff_connect()}}

\item{...}{additional args depending on method}
}
\value{
A tibble
}
\description{
This function returns a tibble with one row for every starter (and bench) for every week and their scoring, if available.
}
