% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfl_transactions.R
\name{ff_transactions.mfl_conn}
\alias{ff_transactions.mfl_conn}
\title{Get full transactions}
\usage{
ff_transactions.mfl_conn()
}
\arguments{
\item{conn}{the list object created by \code{ff_connect()}}
}
\value{
a tibble detailing every transaction since X date
}
\description{
Create a GET request to any MFL export endpoint
}
\examples{
dlf_conn <- mfl_connect(2020,league_id = 37920)
ff_transactions(dlf_conn)

}
