% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfl_userleagues.R
\name{ff_userleagues.mfl_conn}
\alias{ff_userleagues.mfl_conn}
\title{Get User Leagues}
\usage{
\method{ff_userleagues}{mfl_conn}(conn, season = NULL, ...)
}
\arguments{
\item{conn}{a connection object created by \code{ff_connect()}}

\item{season}{the MFL platform season to look for}

\item{...}{arguments that may be passed to other methods (for method consistency)}
}
\description{
This function retrieves a tibble of all user leagues and requires an
authentication cookie to be stored inside the MFL conn object
(which were created via username/password args)
}
