% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ff_userleagues}
\alias{ff_userleagues}
\title{Get User Leagues}
\usage{
ff_userleagues(conn, ...)
}
\arguments{
\item{conn}{a conn object created by \code{ff_connect()}}

\item{...}{additional args which might be used eventually}
}
\value{
A tidy dataframe with one row for every league a user is in
}
\description{
This function returns a tibble with one row for every league a user is in. This requries authentication cookies for MFL usage.
}
