% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffscrapr-package.R
\docType{package}
\name{ffscrapr-package}
\alias{ffscrapr}
\alias{ffscrapr-package}
\title{ffscrapr: API Client for Fantasy Football League Platforms}
\description{
Helps access various Fantasy Football APIs by handling authentication and rate-limiting, forming appropriate calls, and returning tidy dataframes which can be easily connected to other data sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ffscrapr.ffverse.com}
  \item \url{https://github.com/ffverse/ffscrapr}
  \item Report bugs at \url{https://github.com/ffverse/ffscrapr/issues}
}

}
\author{
\strong{Maintainer}: Tan Ho \email{tan@tanho.ca}

Other contributors:
\itemize{
  \item Tony ElHabr \email{anthonyelhabr@gmail.com} [contributor]
  \item Joe Sydlowski \email{syd235@gmail.com} [contributor]
}

}
\keyword{internal}
