% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flea_api.R
\name{fleaflicker_getendpoint}
\alias{fleaflicker_getendpoint}
\title{GET any Fleaflicker endpoint}
\usage{
fleaflicker_getendpoint(endpoint, ...)
}
\arguments{
\item{endpoint}{a string defining which endpoint to return from the API}

\item{...}{Arguments which will be passed as "argumentname = argument" in an HTTP query parameter}
}
\value{
A list object containing the query, response, and parsed content.
}
\description{
The endpoint names and HTTP parameters (i.e. argument names) are CASE SENSITIVE and should be passed in exactly as displayed on the Fleaflicker API reference page.
}
\details{
Check out the vignette for more details and example usage.
}
\seealso{
\url{https://www.fleaflicker.com/api-docs/index.html}

\code{vignette("fleaflicker_getendpoint")}
}
