% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flea_players.R
\name{fleaflicker_players}
\alias{fleaflicker_players}
\title{fleaflicker players library}
\usage{
fleaflicker_players(conn)
}
\value{
a dataframe containing all ~7000+ players in the Fleaflicker database
}
\description{
A cached table of Fleaflicker NFL players. Will store in memory for each session!
(via memoise in zzz.R)
}
\examples{
\donttest{
conn <- fleaflicker_connect(2020,312861)
player_list <- fleaflicker_players(league_id = 312)
}

}
