% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flea_userleagues.R
\name{fleaflicker_userleagues}
\alias{fleaflicker_userleagues}
\title{Fleaflicker - Get User Leagues}
\usage{
fleaflicker_userleagues(user_email, season = NULL)
}
\arguments{
\item{user_email}{the username to look up}

\item{season}{the season to return leagues from - defaults to current year based on heuristics}
}
\value{
a dataframe of leagues for the specified user
}
\description{
This function returns the leagues that a specific user is in.
This variant can be used without first creating a connection object.
}
\seealso{
\code{\link[=ff_userleagues]{ff_userleagues()}}
}
