% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfl.R
\name{get_mfl_endpoint}
\alias{get_mfl_endpoint}
\title{GET any MFL endpoint}
\usage{
get_mfl_endpoint(conn, endpoint, ...)
}
\arguments{
\item{conn}{the list object created by \code{mfl_connect()}}

\item{endpoint}{a string defining which endpoint to return from the API}
}
\value{
the league endpoint for MFL
}
\description{
Create a GET request to any MFL export endpoint, using the parameters defined in \code {mfl_connect()}
}
\examples{
get_mfl_endpoint(conn,endpoint,...)

}
\seealso{
\url {https://api.myfantasyleague.com/2020/api_info?STATE=details}
}
