% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{league_connect}
\alias{league_connect}
\title{Connect to a League}
\usage{
league_connect(platform = "MFL", leagueID, ...)
}
\arguments{
\item{platform}{one of MFL, (future - Sleeper, Flea, ESPN, Yahoo in approximate priority order)}

\item{leagueID}{LeagueID (currently assuming one league at a time)}
}
\value{
a list that stores MFL connection objects
}
\description{
This function creates a connection object which stores parameters and gets a login-cookie if available - it does so by passing arguments to the appropriate league-based handler
}
