% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfl.R
\name{mfl_connect}
\alias{mfl_connect}
\title{Connect to MFL League}
\usage{
mfl_connect(
  season = NULL,
  leagueID = NULL,
  APIKEY = NULL,
  username = NULL,
  password = NULL,
  user_agent = NULL
)
}
\arguments{
\item{season}{Season to access on MFL - if missing, will guess based on system date (current year if March or later, otherwise previous year)}

\item{APIKEY}{APIKEY - optional - allows access to private leagues. Key is unique for each league and accessible from Developer's API page (currently assuming one league at a time)}

\item{username}{MFL username - optional - when supplied in conjunction with a password, will attempt to retrieve authentication token}

\item{password}{MFL password - optional - when supplied in conjunction with username, will attempt to retrieve authentication token}

\item{league}{LeagueID (currently assuming one league at a time)}
}
\value{
a list that stores MFL connection objects
}
\description{
This function creates a connection object which stores parameters and gets a login-cookie if available.
}
