% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfl.R
\name{mfl_endpoint_league}
\alias{mfl_endpoint_league}
\title{Read MFL league settings}
\usage{
mfl_endpoint_league(conn_object)
}
\arguments{
\item{conn_object}{the list object created by \code{mfl_connect()}}
}
\value{
the league endpoint for MFL
}
\description{
Get MFL league settings endpoint based on the predefined terms in mfl_connect()
}
\examples{
mfl_connect(season = 2020, leagueID = 54040) \%>\% mfl_endpoint_league()

}
