% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfl.R
\name{mfl_get_league}
\alias{mfl_get_league}
\title{Read MFL league}
\usage{
mfl_get_league(season = NULL, leagueID = NULL, APIKEY = NULL)
}
\arguments{
\item{season}{Get data for specified league year - if missing, will warn that season is being chosen by sys}

\item{leagueID}{Identifier for league - generally a five digit number found in the homepage URL}

\item{APIKEY}{APIKEY - grants access to views as the user in question, optional for public leagues}
}
\value{
the league endpoint for MFL
}
\description{
Read MFL league
}
\examples{
mfl_get_league(season = 2020, leagueID = 54040, APIKEY = NULL)

}
