% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfl_api.R
\name{mfl_getendpoint}
\alias{mfl_getendpoint}
\title{GET any MFL endpoint}
\usage{
mfl_getendpoint(conn, endpoint, ...)
}
\arguments{
\item{conn}{the list object created by \code{mfl_connect()}}

\item{endpoint}{a string defining which endpoint to return from the API}

\item{...}{Arguments which will be passed as "argumentname = argument" in an HTTP query parameter}
}
\value{
output from the specified MFL API endpoint
}
\description{
Create a GET request to any MFL export endpoint
}
\seealso{
\url{https://api.myfantasyleague.com/2020/api_info?STATE=details}
}
