% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfl.R
\name{mfl_league_summary}
\alias{mfl_league_summary}
\title{Get a summary of common league settings}
\usage{
mfl_league_summary(conn)
}
\arguments{
\item{conn}{the list object created by \code{mfl_connect()}}
}
\value{
the league endpoint for MFL
}
\description{
This function returns a data frame of common league settings - things like "1QB" or "2QB", best ball, team count etc
}
\seealso{
\url{https://api.myfantasyleague.com/2020/api_info?STATE=details}
}
