% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfl_rosters.R
\name{mfl_players}
\alias{mfl_players}
\title{MFL players library}
\usage{
mfl_players()
}
\value{
a dataframe containing all ~2000+ players in the MFL database
}
\description{
A cached table of MFL players. Will store in memory for each session!
(via memoise in zzz.R)
}
\examples{
player_list <- mfl_players()
dplyr::sample_n(player_list,5)

}
