% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfl.R
\name{mfl_scoring_settings}
\alias{mfl_scoring_settings}
\title{Get a dataframe of scoring settings}
\usage{
mfl_scoring_settings(conn)
}
\arguments{
\item{conn}{the list object created by \code{mfl_connect()}}
}
\value{
the league endpoint for MFL
}
\description{
Get a dataframe of scoring settings
}
\seealso{
\url{https://api.myfantasyleague.com/2020/api_info?STATE=details}
}
