% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_deprecated.R
\name{nflfastr_rosters}
\alias{nflfastr_rosters}
\title{Import nflfastr roster data}
\usage{
nflfastr_rosters(...)
}
\arguments{
\item{...}{deprecated}
}
\description{
Deprecated in favour of \code{\link[nflreadr:load_rosters]{nflreadr::load_rosters()}}
}
