% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_helpers.R
\docType{data}
\name{nflfastr_stat_mapping}
\alias{nflfastr_stat_mapping}
\title{Mappings for nflfastr to fantasy platform scoring}
\format{
A data frame with ~85 rows and 3 variables:
\describe{
\item{nflfastr_event}{the column name of the statistic in the nflfastr_weekly dataset}
\item{platform}{specific platform that this mapping applies to}
\item{ff_event}{name of the statistic for that platform}
}
}
\usage{
nflfastr_stat_mapping
}
\description{
A small helper dataframe for connecting nflfastr to specific fantasy platform rules.
}
\keyword{datasets}
