% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_import_nflfastr.R
\name{parse_raw_rds}
\alias{parse_raw_rds}
\title{Parse Raw RDS}
\usage{
parse_raw_rds(raw)
}
\arguments{
\item{raw}{raw-content that is known to be an RDS file}
}
\description{
Useful for parsing the raw-content of RDS files downloaded from nflfastr repo, c/o Seb.
}
\seealso{
\code{httr::set_cookies}
}
\keyword{internal}
