% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_helpers.R
\name{set_unescaped_cookies}
\alias{set_unescaped_cookies}
\title{Add unescaped cookies}
\usage{
set_unescaped_cookies(...)
}
\arguments{
\item{...}{a named cookie values}
}
\description{
Useful for ESPN which is already URL escaped
}
\seealso{
\code{httr::set_cookies}
}
\keyword{internal}
