% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper_draft.R
\name{sleeper_draft}
\alias{sleeper_draft}
\title{Get Sleeper Draft}
\usage{
sleeper_draft(draft_id)
}
\arguments{
\item{draft_id}{draft ID as found in URL e.g. "https://sleeper.com/draft/nfl/{draft_id}"}
}
\value{
draft dataframe
}
\description{
This function retrieves drafts by sleeper's draft ID. This better supports
mock drafts.
}
