% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper_api.R
\name{sleeper_getendpoint}
\alias{sleeper_getendpoint}
\title{GET any SLEEPER endpoint}
\usage{
sleeper_getendpoint(endpoint, ...)
}
\arguments{
\item{endpoint}{a string defining which endpoint to return from the API}

\item{...}{All other arguments are passed in order as slash-separated components of the url}
}
\value{
A list object containing the query, response, and parsed content.
}
\description{
The endpoint names and HTTP parameters (i.e. argument names) are CASE SENSITIVE and should be passed in exactly as displayed on the Sleeper API reference page.
}
\details{
Check out the vignette for more details and example usage.
}
\seealso{
\url{https://docs.sleeper.app}

\code{vignette("sleeper_getendpoint")}
}
