% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper_players.R
\name{sleeper_players}
\alias{sleeper_players}
\title{Sleeper players library}
\usage{
sleeper_players()
}
\value{
a dataframe containing all ~7000+ players in the Sleeper database
}
\description{
A cached table of Sleeper NFL players. Will store in memory for each session!
(via memoise in zzz.R)
}
\examples{
player_list <- sleeper_players()
}
