% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_connect.R
\name{yahoo_connect}
\alias{yahoo_connect}
\title{Connect to Yahoo Fantasy Sports League}
\usage{
yahoo_connect(league_id = NULL, token = NULL, ...)
}
\arguments{
\item{league_id}{League ID}

\item{token}{Token for authentication}
}
\value{
A list that stores Yahoo Fantasy Sports connection objects
}
\description{
This function creates a connection object for Yahoo Fantasy Sports league using the provided league_id and token.
}
\examples{
\donttest{
conn <- yahoo_connect(
  league_id = "12345",
  token = "your_token_here"
)
}

}
