% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_api.R
\name{yahoo_getendpoint}
\alias{yahoo_getendpoint}
\title{GET any Yahoo Fantasy Football API endpoint}
\usage{
yahoo_getendpoint(endpoint, conn)
}
\arguments{
\item{endpoint}{a string defining which endpoint to return from the Yahoo API}

\item{conn}{Yahoo ff_connect object}
}
\value{
A list object containing the query, response, and parsed content as an xml_doc.
}
\description{
GET any Yahoo Fantasy Football API endpoint
}
\seealso{
\url{https://developer.yahoo.com/fantasysports/guide/}
}
