% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper_schedule.R
\name{.sleeper_matchup}
\alias{.sleeper_matchup}
\title{Individual sleeper matchup}
\usage{
.sleeper_matchup(week, conn, last_scored_week)
}
\description{
Individual sleeper matchup
}
\keyword{internal}
