% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_import_dp.R
\name{dp_clean_names}
\alias{dp_clean_names}
\alias{dp_cleannames}
\title{Clean Player Names}
\usage{
dp_clean_names(
  player_name,
  lowercase = FALSE,
  convert_lastfirst = TRUE,
  use_name_database = TRUE
)

dp_cleannames(
  player_name,
  lowercase = FALSE,
  convert_lastfirst = TRUE,
  use_name_database = TRUE
)
}
\value{
a character vector of cleaned names
}
\description{
Clean Player Names
}
\examples{
\donttest{

dp_cleannames(c("A.J. Green", "Odell Beckham Jr.", "Le'Veon Bell Sr."))

dp_cleannames(c("Trubisky, Mitch", "Atwell, Chatarius", "Elliott, Zeke", "Elijah Moore"),
  convert_lastfirst = TRUE,
  use_name_database = TRUE
)
}

}
\seealso{
\code{nflreadr::clean_player_names()}
}
