% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_import_dp.R
\docType{data}
\name{dp_name_mapping}
\alias{dp_name_mapping}
\title{Alternate name mappings}
\format{
A named character vector
\describe{
\item{name attribute}{The "alternate" name.}
\item{value attribute}{The "correct" name.}
}
}
\usage{
dp_name_mapping
}
\description{
A named character vector mapping common alternate names
}
\examples{
\donttest{
dp_name_mapping[c("Chatarius Atwell", "Robert Kelley")]
}

}
\keyword{datasets}
