% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_import_dp.R
\name{dp_playerids}
\alias{dp_playerids}
\title{Import latest DynastyProcess player IDs}
\usage{
dp_playerids()
}
\value{
a tibble of player IDs
}
\description{
Fetches a copy of the latest DynastyProcess player IDs csv
}
\examples{
\donttest{
try( # try only shown here because sometimes CRAN checks are weird
  dp_playerids()
)
}

}
\seealso{
\url{https://github.com/DynastyProcess/data}
}
