% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_import_dp.R
\name{dp_values}
\alias{dp_values}
\title{Import latest DynastyProcess values}
\usage{
dp_values(file = c("values.csv", "values-players.csv", "values-picks.csv"))
}
\arguments{
\item{file}{one of \code{c("values.csv","values-players.csv","values-picks.csv")}}
}
\value{
a tibble of trade values from DynastyProcess
}
\description{
Fetches a copy of the latest DynastyProcess dynasty trade values sheets
}
\examples{
\donttest{
try( # try only shown here because sometimes CRAN checks are weird
  dp_values()
)
}

}
\seealso{
\url{https://github.com/DynastyProcess/data}
}
