% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_api.R
\name{espn_getendpoint}
\alias{espn_getendpoint}
\title{GET any ESPN endpoint}
\usage{
espn_getendpoint(conn, x_fantasy_filter = NULL, ...)
}
\arguments{
\item{conn}{a connection object created by \code{espn_connect} or \code{ff_connect()}}

\item{x_fantasy_filter}{a JSON-encoded character string that specifies a filter for the data}

\item{...}{Arguments which will be passed as "argumentname = argument" in an HTTP query parameter}

\item{url}{a fully formed URL call, will override all other URL components}
}
\value{
A list object containing the query, response, and parsed content.
}
\description{
The endpoint names and HTTP parameters (i.e. argument names) are CASE SENSITIVE.
Best URL help page is TBD, possibly might just be the vignette.
}
\details{
Check out the vignette for more details and example usage.
}
\seealso{
\code{vignette("espn_getendpoint")}
}
