% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_api.R
\name{espn_getendpoint}
\alias{espn_getendpoint}
\title{GET ESPN fantasy league endpoint}
\usage{
espn_getendpoint(conn, ..., x_fantasy_filter = NULL)
}
\arguments{
\item{conn}{a connection object created by \code{espn_connect} or \code{ff_connect()}}

\item{...}{Arguments which will be passed as "argumentname = argument" in an HTTP query parameter}

\item{x_fantasy_filter}{a JSON-encoded character string that specifies a filter for the data}
}
\value{
A list object containing the query, response, and parsed content.
}
\description{
This function is used to call the ESPN Fantasy API for league-based endpoints.
}
\details{
The ESPN Fantasy API is undocumented and this should be used by advanced users
familiar with the API.

It chooses the correct league endpoint based on the year (eg leagueHistory
for <2018), checks the x_fantasy_filter for valid JSON input, builds a url
with any optional query parameters, and executes the request with authentication
and rate limiting.

HTTP query parameters (i.e. arguments to ...) are Case Sensitive.

Please see the vignette for more on usage.
}
\seealso{
\code{vignette("espn_getendpoint")}

\code{espn_getendpoint_raw}
}
