% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_players.R
\name{espn_players}
\alias{espn_players}
\title{ESPN players library}
\usage{
espn_players(conn = NULL, season = NULL)
}
\arguments{
\item{conn}{a connection object created by \code{espn_connect} or \code{ff_connect()}}

\item{season}{a season to fetch}
}
\value{
a dataframe containing all ~2000+ active players in the ESPN database
}
\description{
A cached table of ESPN NFL players. Will store in memory for each session!
(via memoise in zzz.R)
}
\examples{
\donttest{
try({ # try only shown here because sometimes CRAN checks are weird

  conn <- espn_connect(season = 2020, league_id = 1178049)

  espn_players(conn, season = 2020)
}) # end try
}

}
