% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mfl_draft.R
\name{ff_draft}
\alias{ff_draft}
\alias{ff_draft.mfl_conn}
\title{Get Draft Results}
\usage{
ff_draft(conn, ...)

\method{ff_draft}{mfl_conn}(conn, custom_players = FALSE, ...)
}
\arguments{
\item{conn}{a conn object created by \code{ff_connect()}}

\item{...}{not sure if there'll be other params yet!}

\item{custom_players}{TRUE or FALSE - retrieve custom players from the MFL database? (Devy, placeholder picks etc)}
}
\value{
A tibble of draft results
}
\description{
This function gets a table of the draft results for the current year.
Can handle MFL devy drafts or startup drafts by specifying the custom_players argument
}
\examples{
ssb_conn <- ff_connect(platform = "mfl",league_id = 54040,season = 2020)
ff_draft(ssb_conn)

}
