% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mfl_draftpicks.R
\name{ff_draftpicks}
\alias{ff_draftpicks}
\alias{ff_draftpicks.mfl_conn}
\title{Get Draft Picks}
\usage{
ff_draftpicks(conn, ...)

\method{ff_draftpicks}{mfl_conn}(conn, ...)
}
\arguments{
\item{conn}{the list object created by \code{ff_connect()}}

\item{...}{other arguments (currently unused)}
}
\value{
A tibble of draft picks

a tibble detailing future draft picks
}
\description{
This function returns a tibble of current and future draft picks

Get current and future draft picks that have not yet been selected/converted into players
}
\examples{
dlf_conn <- mfl_connect(2020,league_id = 37920)
ff_draftpicks(conn = dlf_conn)

}
