% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mfl_draftpicks.R,
%   R/sleeper_draftpicks.R
\name{ff_draftpicks}
\alias{ff_draftpicks}
\alias{ff_draftpicks.mfl_conn}
\alias{ff_draftpicks.sleeper_conn}
\title{Get Draft Picks}
\usage{
ff_draftpicks(conn, ...)

\method{ff_draftpicks}{mfl_conn}(conn, ...)

\method{ff_draftpicks}{sleeper_conn}(conn, ...)
}
\arguments{
\item{conn}{the list object created by \code{ff_connect()}}

\item{...}{other arguments (currently unused)}
}
\value{
a tibble detailing future draft picks

a tibble detailing future draft picks
}
\description{
Get current and future draft picks that have not yet been selected/converted into players

Get current and future draft picks that have not yet been selected/converted into players
}
\section{Methods (by class)}{
\itemize{
\item \code{mfl_conn}: MFL: returns both current and future picks

\item \code{sleeper_conn}: Sleeper: retrieves current and future draft picks
}}

\examples{
dlf_conn <- mfl_connect(2020, league_id = 37920)
ff_draftpicks(conn = dlf_conn)
jml_conn <- ff_connect(platform = "sleeper", league_id = 522458773317046272, season = 2020)
ff_draftpicks(jml_conn)
}
