% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mfl_franchises.R
\name{ff_franchises}
\alias{ff_franchises}
\alias{ff_franchises.mfl_conn}
\title{Get League Franchises}
\usage{
ff_franchises(conn)

ff_franchises.mfl_conn(conn)
}
\arguments{
\item{conn}{a conn object created by \code{ff_connect()}}
}
\value{
A tibble of franchises
}
\description{
This function returns a tibble of franchise details (id, name, abbrev etc)

Returns a tidy dataframe of franchises and any relevant information
}
\examples{
ssb_conn <- ff_connect(platform = "mfl",league_id = 54040,season = 2020)
ff_franchises(ssb_conn)

}
