% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mfl_league.R
\name{ff_league}
\alias{ff_league}
\alias{ff_league.mfl_conn}
\title{Get League Summary}
\usage{
ff_league(conn)

\method{ff_league}{mfl_conn}(conn)
}
\arguments{
\item{conn}{the connection object created by \code{mfl_connect()}}
}
\value{
A one-row tibble of scoring settings.
}
\description{
This function returns a dataframe summarizing common league settings.

This function returns a tibble of common league settings - things like "1QB" or "2QB", best ball, team count etc
}
\examples{
ff_connect(platform = "mfl",league_id = 54040,season = 2020) \%>\% ff_league() \%>\% str()

}
