% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_generics.R, R/flea_scoringhistory.R,
%   R/mfl_scoringhistory.R, R/sleeper_scoringhistory.R
\name{ff_scoringhistory}
\alias{ff_scoringhistory}
\alias{ff_scoringhistory.flea_conn}
\alias{ff_scoringhistory.mfl_conn}
\alias{ff_scoringhistory.sleeper_conn}
\title{Get a dataframe of scoring history, utilizing the ff_scoring and load_player_stats functions.}
\usage{
ff_scoringhistory(conn, season, ...)

\method{ff_scoringhistory}{flea_conn}(conn, season = 1999:2020, ...)

\method{ff_scoringhistory}{mfl_conn}(conn, season = 1999:2020, ...)

\method{ff_scoringhistory}{sleeper_conn}(conn, season = 1999:2020, ...)
}
\arguments{
\item{conn}{a conn object created by \code{ff_connect()}}

\item{season}{season a numeric vector of seasons (earliest available year is 1999)}

\item{...}{other arguments}
}
\value{
A tidy dataframe of scoring history in a flat table, one row per player per week.
}
\description{
Get a dataframe of scoring history, utilizing the ff_scoring and load_player_stats functions.

Get a dataframe of scoring history, utilizing the ff_scoring and load_player_stats functions.

Get a dataframe of scoring history, utilizing the ff_scoring and load_player_stats functions.

Get a dataframe of scoring history, utilizing the ff_scoring and load_player_stats functions.
}
\section{Methods (by class)}{
\itemize{
\item \code{flea_conn}: Fleaflicker: returns scoring history in a flat table, one row per player per week.

\item \code{mfl_conn}: MFL: returns scoring history in a flat table, one row per player per week.

\item \code{sleeper_conn}: Sleeper: returns scoring history in a flat table, one row per player per week.
}}

\examples{
\donttest{#'
conn <- fleaflicker_connect(2020, 312861)
x <- ff_scoringhistory(conn, season = 2020)
x
}

\donttest{
}

\donttest{#'
conn <- ff_connect(platform = "sleeper", league_id = "522458773317046272", season = 2020)
ff_scoringhistory(conn, season = 2020)
}

}
\seealso{
\url{https://www.nflfastr.com/reference/load_player_stats.html}

\url{https://www.nflfastr.com/reference/load_player_stats.html}

\url{https://www.nflfastr.com/reference/load_player_stats.html}
}
