% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mfl_standings.R,
%   R/sleeper_standings.R
\name{ff_standings}
\alias{ff_standings}
\alias{ff_standings.mfl_conn}
\alias{ff_standings.sleeper_conn}
\title{Get Standings}
\usage{
ff_standings(conn, ...)

\method{ff_standings}{mfl_conn}(conn, ...)

\method{ff_standings}{sleeper_conn}(conn, ...)
}
\arguments{
\item{conn}{a conn object created by \code{ff_connect()}}

\item{...}{arguments passed to other methods (currently none)}
}
\value{
A tibble of standings
}
\description{
This function returns a tibble of season-long fantasy team stats
}
\examples{
ssb_conn <- ff_connect(platform = "mfl", league_id = 54040, season = 2020)
ff_standings(ssb_conn)
jml_conn <- ff_connect(platform = "sleeper", league_id = 522458773317046272, season = 2020)
ff_standings(jml_conn)
}
