% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mfl_transactions.R
\name{ff_transactions}
\alias{ff_transactions}
\alias{ff_transactions.mfl_conn}
\title{Get League Transactions}
\usage{
ff_transactions(conn, ...)

\method{ff_transactions}{mfl_conn}(conn, custom_players = FALSE, ...)
}
\arguments{
\item{conn}{the list object created by \code{ff_connect()}}

\item{...}{additional args}

\item{custom_players}{TRUE or FALSE - fetch custom players}
}
\value{
A tidy dataframe of transaction data
}
\description{
This function returns a tibble of transactions
}
\examples{
dlf_conn <- mfl_connect(2019, league_id = 37920)
ff_transactions(dlf_conn)
}
