% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mfl_userleagues.R,
%   R/sleeper_userleagues.R
\name{ff_userleagues}
\alias{ff_userleagues}
\alias{ff_userleagues.mfl_conn}
\alias{ff_userleagues.sleeper_conn}
\title{Get User Leagues}
\usage{
ff_userleagues(conn, ...)

\method{ff_userleagues}{mfl_conn}(conn, season = NULL, ...)

\method{ff_userleagues}{sleeper_conn}(conn, user_name = NULL, ...)
}
\arguments{
\item{conn}{a connection object created by \code{ff_connect()}}

\item{...}{arguments that may be passed to other methods (for method consistency)}

\item{season}{the MFL platform season to look for}

\item{user_name}{the username to look up - defaults to user created in conn if available}
}
\value{
A tidy dataframe with one row for every league a user is in
}
\description{
This function returns a tibble with one row for every league a user is in. This requries authentication cookies for MFL usage.

This function retrieves a tibble of all user leagues and requires an
authentication cookie to be stored inside the MFL conn object
(which were created via username/password args)

This function retrieves a tibble of all user leagues
(which were created via username/password args)
}
\section{Methods (by class)}{
\itemize{
\item \code{mfl_conn}: MFL: With an authentication token, it will return a list of user leagues.

\item \code{sleeper_conn}: Sleeper: returns a listing of leagues for a given user_id or user_name
}}

