% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flea_players.R
\name{fleaflicker_players}
\alias{fleaflicker_players}
\title{Fleaflicker players library}
\usage{
fleaflicker_players(conn, page_limit = NULL)
}
\arguments{
\item{conn}{a conn object created by \code{ff_connect()}}

\item{page_limit}{A number limiting the number of players to return, or NULL (default) returns all}
}
\value{
a dataframe containing all ~7000+ players in the Fleaflicker database
}
\description{
A cached table of Fleaflicker NFL players. Will store in memory for each session!
(via memoise in zzz.R)
}
\examples{
\donttest{
try({ # try only shown here because sometimes CRAN checks are weird
  conn <- fleaflicker_connect(2020, 312861)
  player_list <- fleaflicker_players(conn, page_limit = 2)
}) # end try
}

}
