% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_import_nflverse.R
\name{nflfastr_rosters}
\alias{nflfastr_rosters}
\title{Import nflfastr roster data}
\usage{
nflfastr_rosters(seasons)
}
\description{
Deprecated in favour of \code{\link[nflreadr:load_rosters]{nflreadr::load_rosters()}}
}
