% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_deprecated.R
\name{nflfastr_weekly}
\alias{nflfastr_weekly}
\title{Import latest nflfastr weekly stats}
\usage{
nflfastr_weekly(...)
}
\description{
Deprecated in favour of \code{\link[nflreadr:load_player_stats]{nflreadr::load_player_stats()}}
}
