% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper.R
\name{sleeper_getendpoint}
\alias{sleeper_getendpoint}
\title{GET Sleeper endpoint
## NOT YET WORKING ##}
\usage{
sleeper_getendpoint(conn, endpoint, ...)
}
\arguments{
\item{conn}{the list object created by \code{mfl_connect()}}

\item{endpoint}{a string defining which endpoint to return from the API}

\item{...}{Arguments which will be passed as "argumentname = argument" in an HTTP query parameter}
}
\value{
the league endpoint for MFL
}
\description{
Create a GET request to any MFL export endpoint
}
\seealso{
\url{https://api.myfantasyleague.com/2020/api_info?STATE=details}
}
