% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yahoo_api.R
\name{yahoo_getendpoint}
\alias{yahoo_getendpoint}
\title{GET any Yahoo Fantasy Football API endpoint}
\usage{
yahoo_getendpoint(conn, endpoint)
}
\arguments{
\item{endpoint}{a string defining which endpoint to return from the Yahoo API}

\item{token}{Yahoo Fantasy Football API token for authentication}

\item{...}{Arguments which will be passed as "argumentname = argument" in an HTTP query parameter}
}
\value{
A list object containing the query, response, and parsed content.
}
\description{
The endpoint names and HTTP parameters (i.e., argument names) are CASE SENSITIVE and should be passed in exactly as required by the Yahoo Fantasy Football API.
}
